fix: rm useless comments and move clean deprecated methods
